% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/False_Positive.R
\name{false_positive}
\alias{false_positive}
\title{False Positives}
\usage{
false_positive(P = P, QTN.position, cutoff)
}
\arguments{
\item{P}{P values obtained from GWAS}

\item{QTN.position}{Indexes showing the QTN position in theGenotypic Matrix}

\item{cutoff}{Threshold/Significance level to assess significance}
}
\value{
Proportion of false positives
}
\description{
Calculates the proportion of False Positives
}
