% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan Plot}
\usage{
manhattan_plot(data, cutoff, p, QTN.position = NULL)
}
\arguments{
\item{data}{the individual and marker information (marker map file with m by 3 dimension where m is number of markers in row)}

\item{cutoff}{threshold for identifying the significant SNPs}

\item{p}{p-value of each markers}

\item{QTN.position}{the position of the significant QTN}
}
\value{
plots and graphs
}
\description{
Function to generate Manhattan plot for visual display of significant markers
}
