% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-value-GLM.R
\name{p_val_GLM}
\alias{p_val_GLM}
\title{Function to return p-value for SNP}
\usage{
p_val_GLM(y, X, C = NULL)
}
\arguments{
\item{y}{Phenotype matrix in the form n by 1 (n is number of individual in rows)}

\item{X}{Marker data in matrix form n by m ( n is number of individual in row and m is marker data in column)}

\item{C}{Covariates matrix in the form n by t (n is number of individuals in row and t is number of covariates in column)}
}
\description{
This is the function to calculate and return p-values for each SNPs after testing the association between the Phenotypes and SNP along with their Covairates
}
