% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PC}
\alias{PC}
\title{Principle Component Analysis}
\usage{
PC(PC_no, X, C)
}
\arguments{
\item{PC_no}{User specified number of Principle Components (PCs)}

\item{X}{Markers data in the form n by m with n number of individuals and m number of markers}

\item{C}{Covariates matrix in the form n by t with n number of individuals and t number of coavariates}
}
\value{
Principle Component Analysis
}
\description{
Removes PCs that are linearly dependent with the given covariates and also user can specify how many PCs to chooose as co-factors
}
