% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_data}
\alias{import_data}
\title{Function to import data file}
\format{matrix}
\usage{
import_data(filename)
}
\arguments{
\item{filename}{Input file to import (Requires Phenotype, genotype, map and covariates information)}
}
\value{
data file
}
\description{
A function to import data file. The import data contains information about Phenotype (n by 1 form), Markers (Numeric as well as map file with n by m and m by 3 dimension respectively) and Covariates (n by t form) with n the number of individuals in row, m number of markers in column and t number of covariates in column
}
